requirements.txt
django
django-recaptcha
Pillow
mysqlclient
user-agents
djangorestframework
requests